+++ nix/simit-module.nix
@@ -1,5 +1,83 @@
 {lib}: {
   options.simit = {
+    flake = lib.mkOption {
+      default = {};
+      type = lib.types.submodule {
+        freeformType = lib.types.attrsOf lib.types.anything;
+        options = {
+          mode = lib.mkOption {
+            default = "canonical";
+            type = lib.types.enum ["canonical" "custom"];
+          };
+          toolchain_binding = lib.mkOption {
+            default = "rustToolchain";
+            type = lib.types.str;
+          };
+          crane_lib_binding = lib.mkOption {
+            default = "craneLib";
+            type = lib.types.str;
+          };
+          package_binding = lib.mkOption {
+            default = "package";
+            type = lib.types.str;
+          };
+          formatter_output = lib.mkOption {
+            default = true;
+            type = lib.types.bool;
+          };
+          formatting_check = lib.mkOption {
+            default = true;
+            type = lib.types.bool;
+          };
+          pre_commit_shell_hook = lib.mkOption {
+            default = true;
+            type = lib.types.bool;
+          };
+          expected_outputs = lib.mkOption {
+            default = {};
+            type = lib.types.submodule {
+              freeformType = lib.types.attrsOf lib.types.anything;
+              options = {
+                packages = lib.mkOption {
+                  default = [];
+                  type = lib.types.listOf lib.types.str;
+                };
+                checks = lib.mkOption {
+                  default = [];
+                  type = lib.types.listOf lib.types.str;
+                };
+                top_level = lib.mkOption {
+                  default = [];
+                  type = lib.types.listOf lib.types.str;
+                };
+              };
+            };
+          };
+        };
+      };
+    };
+
+    ci = lib.mkOption {
+      default = {};
+      type = lib.types.submodule {
+        freeformType = lib.types.attrsOf lib.types.anything;
+        options = {
+          extra_setup = lib.mkOption {
+            default = [];
+            type = lib.types.listOf lib.types.str;
+          };
+          extra_env = lib.mkOption {
+            default = {};
+            type = lib.types.attrsOf lib.types.str;
+          };
+          required_secrets = lib.mkOption {
+            default = [];
+            type = lib.types.listOf lib.types.str;
+          };
+        };
+      };
+    };
+
     release = lib.mkOption {
       default = {};
       type = lib.types.submodule {